move interfaces and enums to separate folders
