+++ src/app/components/tables/data-table/enums/terra-ref-type.enum.ts
@@ -1,0 +1,7 @@
+export enum TerraRefTypeEnum
+{
+    email = 'mailto',
+    phone = 'tel',
+    url = 'url',
+    function = 'function'
+}
