+++ src/app/components/tables/data-table/enums/terra-text-align.enum.ts
