+++ src/app/components/tables/data-table/example/terra-data-table.component.example.ts
@@ -3,10 +3,10 @@ import {
     OnInit
 } from '@angular/core';
 import { TerraDataTableServiceExample } from './terra-data-table.service.example';
-import { TerraDataTableRowInterface } from '../row/terra-data-table-row.interface';
-import { TerraDataTableHeaderCellInterface } from '../cell/terra-data-table-header-cell.interface';
-import { TerraTextAlignEnum } from '../cell/terra-text-align.enum';
-import { TerraDataTableCellInterface } from '../cell/terra-data-table-cell.interface';
+import { TerraDataTableRowInterface } from '../interfaces/terra-data-table-row.interface';
+import { TerraDataTableHeaderCellInterface } from '../interfaces/terra-data-table-header-cell.interface';
+import { TerraTextAlignEnum } from '../enums/terra-text-align.enum';
+import { TerraDataTableCellInterface } from '../interfaces/terra-data-table-cell.interface';
 import { TerraButtonInterface } from '../../../buttons/button/data/terra-button.interface';
 
 @Component({
