+++ src/app/components/tables/data-table/interfaces/terra-data-table-cell.interface.ts
