+++ src/app/components/tables/data-table/interfaces/terra-data-table-text.interface.ts
