+++ src/index.ts
@@ -31,10 +31,10 @@ export * from './app/components/pager/data/terra-pager.interface';
 export * from './app/components/pager/terra-pager.component';
 export * from './app/components/layouts/portlet/terra-portlet.component';
 export * from './app/service/terra-base.service';
-export * from './app/components/tables/data-table/cell/terra-data-table-cell.interface';
-export * from './app/components/tables/data-table/cell/terra-ref-type.interface';
-export * from './app/components/tables/data-table/cell/terra-data-table-text.interface';
-export * from './app/components/tables/data-table/cell/terra-data-table-header-cell.interface';
+export * from './app/components/tables/data-table/interfaces/terra-data-table-cell.interface';
+export * from './app/components/tables/data-table/interfaces/terra-ref-type.interface';
+export * from './app/components/tables/data-table/interfaces/terra-data-table-text.interface';
+export * from './app/components/tables/data-table/interfaces/terra-data-table-header-cell.interface';
 export * from './app/components/tables/simple/cell/terra-simple-table-cell.interface';
 export * from './app/components/tables/simple/cell/terra-simple-table-header-cell.interface';
 export * from './app/components/tables/data-table/context-menu/directive/terra-data-table-context-menu.directive';
@@ -42,7 +42,7 @@ export * from './app/components/tables/data-table/context-menu/data/terra-data-t
 export * from './app/components/tables/data-table/context-menu/service/terra-data-table-context-menu.service';
 export * from './app/components/tables/data-table/context-menu/terra-data-table-context-menu.component';
 export * from './app/components/tables/data-table/terra-data-table.component';
-export * from './app/components/tables/data-table/row/terra-data-table-row.interface';
+export * from './app/components/tables/data-table/interfaces/terra-data-table-row.interface';
 export * from './app/components/tables/simple/terra-simple-table.component';
 export * from './app/components/tables/simple/row/terra-simple-table-row.interface';
 export * from './app/components/layouts/tag/terra-tag.component';
@@ -82,7 +82,7 @@ export * from './app/components/tree/node-tree/data/terra-node.interface';
 export * from './app/components/tree/node-tree/node/terra-node.component';
 export * from './app/components/forms/input/time-picker/terra-time-picker.component';
 export * from './app/components/dynamic-module-loader/data/terra-dynamic-loaded-component-input.interface';
-export * from './app/components/tables/data-table/cell/terra-text-align.enum';
+export * from './app/components/tables/data-table/enums/terra-text-align.enum';
 export * from './app/components/data/terra-base-parameter.interface';
 export * from './app/components/forms/dynamic-form/terra-dynamic-form.component';
 export * from './app/components/forms/dynamic-form/dynamic-switch/terra-dynamic-switch.component';
